Have Elem native helpers return TypedValue
Summary:
The JIT versions of the Elem helpers already return value types; this change simply makes the native helpers return the same types. To handle the new type in the interpreter, we do the same thing that we're doing in the JIT: we store the value to tvTempBase (which doesn't require refcounting) and get an lval to it.

This change allows a few nice cleanups:
- The ElemX IR op no longer needs to take tvRef, since it can just return a new value.
- We can eliminate the ugly ArrayData::rvalVanilla now that ArrayData::get satisfies all use cases.
- We can eliminate the tv_rval::as_lval() "const-cast".

As follow-ups to this change:
- I'll modify NvGetInt / NvGetStr to do ArrayData::get directly, which should cut out a (non-perf-affecting) branch added in D21197185.
- I'll use tvTempBase for the define / unset case so we can eliminate the idea of "ratcheting" tvRef and tvRef2.

Reviewed By: ricklavoie

Differential Revision: D21257230

fbshipit-source-id: 785c23ad4ecca74aad4ec9c1e74413edc4e34570