De-virtualize ArrayIdx and ArrayGet
Summary:
I'm re-landing this change now that I've collected more perf data. It's flat in production load tests, but helps with the Hack Array Migration (because we have much more layout information when we start checking darray and varray typehints). HAM specialization is flat with this change.

Modify the DictIdx and DictGet native helpers to handle mixed arrays as well, then dispatch to those helpers if we have a known TMixedArr input for ArrayIdx or ArrayGet. It's better to share these native helpers because these functions are all very hot so we can share the text itself.

arrayIdxS and arrayGetS have the same signature as dictIdxS and dictGetS, and they call them in tail-position, so the array functions should just jump into the dict functions and BOLT should arrange them right next to each other. I've confirmed that that happens with objdump.

In addition, simplify ArrayIdx and DictIdx to ArrayGet<None> and DictGetQuiet if their default argument is InitNull - by doing so, we avoid setting up two ABI registers.

Reviewed By: jjgriego

Differential Revision: D21280797

fbshipit-source-id: dce9880ee7725b11a1f22e1f3c2ad55c1bee3dc7