Some tracing fixes/improvements
Summary:
Some miscellaneous tracing fixes/improvements. Right now, if the
tracing backend decides it wants to trace, we bypass all the coin-flip
logic and always trace. This is somewhat confusing, and makes it hard
to estimate the amount of logging we're done. Instead, always do the
coin-flip, and only ask the tracing backend if we the coin-flip
succeeds. Since we're always doing the coin-flip, the notion of a
"sample rate" is meaningful, so report that via StructuredLog.

Differential Revision: D21273628

fbshipit-source-id: 9ddc8c85663ae1ff715a7107306203bdcdb705ef