Stop considering failed tests as passed or skipped in server mode.
Summary:
We had weird logic that, in server mode, would re-run failed tests in standalone mode and either show them as passed (if the the curl command in server mode had failed but the test passed in standalone mode) or skipped (if the test failed some other way but passed in standalone mode). Stop doing this.

Also remove bizzare logic that was returning 'pass-server' and 'skip-server' (rather than true) as a passing results so that when we don't have cursor control but we're still being run interactively we'd output a red dot and/or a green comma instead of a green dot for the "passed" tests (which still counted as passed in the overall test run regardless).

Fix one test that was using the STDIN constant unavailable in server mode to use a different file resource that is available in all modes.

There are now 171 slow tests that (correctly) fail in server mode (and were previously being marked skipped).

Reviewed By: markw65

Differential Revision: D21244371

fbshipit-source-id: fc39ff20d6d0283a5a53c66892a8f84ceaea16c4