Use Criterion for direct-decl-parser benchmark
Summary:
Julia wrote this parser benchmark program using rustc's libtest, which provides some very simple benchmarking facilities. It's useful for getting a very coarse picture of how the parsers compare in performance, but I found it very difficult to determine the performance impact of a change because of how noisy the benchmark's results tended to be. I tried to find a way to increase the number of iterations run by the benchmark, but libtest does not seem to provide one, so I investigated other benchmarking libraries.

Criterion seems to be the leading library at the moment externally. It's used by Mononoke in fbcode as well (D20720789). Although benchmark runs on devservers are still very noisy, I've found that being able to adjust the measurement time helps me control that noise somewhat.

Reviewed By: arxanas

Differential Revision: D21145262

fbshipit-source-id: a66ea5af79550860e8ac840cc4ea2cc817097e6c