fix getFile for Throwable
Summary:
We noticed this problem when debugging some production problems. Essentially if one calls `$ex->getFile()` within a trait, the call will wrongly attribute to the file of the class using the trait, not the file of the trait itself.

There are 2 places this needs fixing but the fix is essentially going to check the `originalFilename` and then fallback to `unit()->file()` and this behavior is already in `$function->filename()`.

Reviewed By: mofarrell

Differential Revision: D21303631

fbshipit-source-id: 6ca110b83c48eeb2e18c384e271e025c8ae05b43