Fix ClsMeth dvarray typehint behavior
Summary:
ClsMeth may escalate to a varray when tested against certain typehints. Escalation is tricky because we need to emit any dvarray notices or checks when we do so. There are four places where we escalate ClsMeth to varray. They're the cross product {interp, JIT} x {Hack typehint, builtin typehint}:
- Interp Hack typehints: type-constraint.cpp
- Interp native typehints: native.cpp
- JIT Hack typehints: irgen-types.cpp
- JIT native typehints: irgen-builtin.cpp

In this diff, we consolidate and fix the logic in these four places. The core logic for whether we can escalate a ClsMeth for a given typehint is in annotCompat; we add a convertClsMethToArrLike helper to TypeConstraint that's a thin convenience shim around this method. Consolidating this logic lets modify a single site for dvarray specialization.

We also add a raiseClsMethHackArrCompatNotice to determine whether, post escalation, we need a notice. Previously, all four places only raised a notice when we escalated for a darray typehint, but we also need to handle the plain array case, and, when the TypeHintPolymorphism flag is on, the varray_or_darray case. Now we do all of them.

The flow in all four sites is now:
- Call convertClsMethToArrLike to determine whether we should escalate
- Emit the IsVec notice noting the escalation
- Do the escalation
- Call raiseClsMethHackArrCompatNotice and maybe emit a HackArrCompat notice

Reviewed By: paulbiss

Differential Revision: D21297228

fbshipit-source-id: 18efc061355008a638bccc2349d07e3219adef1e