Remove the TypeHintPolymorphism flag
Summary:
Remove this flag and perform DCE.

We may want to bring this flag back at some point to clean up varray_or_darray typehints in the codebase, but we'll probably also want it to apply to vec_and_dict if we do so. viratyosin suggests that we may be able to introduce a bounded soft typehint for this purpose instead, which would solve the problem more generally.

Reviewed By: hgoldstein

Differential Revision: D21315300

fbshipit-source-id: 869df5dce624081c4eab19f904240fe8c0f5c8ed