Fix array refinement bug
Summary:
The refinement in `is_array` tests does the wrong thing when the static type of the variable under inspection is not concrete e.g. a type parameter. Fix is to add implied constraints only on concrete types, as with `is` tests.

Future: we should unify the code for `is` and `is_array`, `is_dict`, etc tests. There is a lot of commonality.

Reviewed By: CatherineGasnier

Differential Revision: D21282430

fbshipit-source-id: 3e55e86b09bbb819e19d47d1e0c802924ab747bd