Don't leak m_setprofileFunctions
Summary:
g_context is normally swept, not destroyed, so neither the hash-set,
nor its contents are destroyed at the end of a request. Fortunately,
the set is not normally used (and hence holds no memory).

We can use a req::hash_set<String> instead, which also avoids
conversions back and forth when constructing, and inspecting the
set. While we're here, make the set case insensitive, to match the
runtime behavior.

Reviewed By: oulgen

Differential Revision: D21306975

fbshipit-source-id: 404901492855e75d2819e43efea45fd241e46ad5