Remove support for __get, __set, __isset and __unset from parser
Summary:
Make it a parse error to define a method named __get, __set, __isset or __unset.

Add new tests to the typechecker and HHVM that verify that the error is
correctly produced. Remove all obsoleted tests and update those where the
primary purpose was not to test magic methods.

The next diff will remove handling of these methods in HHVM.

Differential Revision: D21294504

fbshipit-source-id: 6fce756397a3eaf2d4181b627e88ae987446414f