Get rid of tvRef for remaining Elem helpers
Summary:
All of the remaining usages of tvRef / tvRef2 for Elem helpers can be replaced with either a) handling out an lval to immutable_null_base or b) returning a value instead. We modify ElemDX and ElemUX to do a) and SetOpElem and SetOpNewElem to do b).

Of course, a) is sketchy, since we had better not modify immutable_null_base! We've already been using this trick in a number of places, though. I added assertions in ElemD and ElemU to alert us to invalid modifications of this sentinel.

Reviewed By: jano

Differential Revision: D21328409

fbshipit-source-id: b1fa98118e646b2a03b10a4b2a13f83bd7e946e5