Do not check wrong class when there are duplicate definitions of the class
Summary: The code in `Typing_toplevel.class_def` checks a class having obtained "the decl" for the class using `Typing_env.get_class`. But it might be the wrong decl, if the class has duplicate definitions! Naming has already produced an error for duplicates, but we should not proceed to check a class wrt the wrong decl.

Reviewed By: vsiles

Differential Revision: D21301422

fbshipit-source-id: 7257eeb928fddc731c12eb0ae1e44d84f1f5cf34