Remove XML UseSafeSubtrees - it's always true
Summary:
When the UseSafeSubtrees flag is off, the order of XML node destruction is visible to callers because destroying the root node will actually destroy child resources. (We don't crash in this case, because the accessors of those resources are native property handlers that check for deletion and throw an error.) This behavior is weird and doesn't get used in any production configuration that I can see - the UseSafeSubtrees flag defaults to true and is only overridden in tests. Remove it and do DCE.

For tests, there are a few tests that specifically depend on this behavior (both flag on and off). I removed the subdirectory for "safe-trees" tests and modified the default tests to assume that the flag is on instead.

This behavior is blocking for eliminating MInstr tvRef / tvRef2, which we'd like to do to simplify that code.

Reviewed By: paulbiss, jano

Differential Revision: D21338422

fbshipit-source-id: 6df499a2d78c19c2862739ff650a87af6f8a9c09