tell user why saved-state failed
Summary:
This is just boring plumbing. When an external tool failed during saved-state-init, we should expose the error text from that tool directly to the user.

Basically, the diff is just about taking care to (1) have variables that clearly indicate what is user-facing and what is destined only for our logs, (2) thread them through.

Reviewed By: 2BitSalute

Differential Revision: D21343546

fbshipit-source-id: 20bd4c1916f4a0ece3e2d81cfbc5bb27221bf9a6