Support dvarrays in builtin error messages
Summary:
We use ThrowParameterWrongType to raise param type error messages for builtins, rather than VerifyParamFail. This IR op raises a different error type and message, and has semantic implications (e.g. the builtin error is catchable but VerifyParamFail fatals) that we don't want to mess with right now.

ThrowParameterWrongType takes the expected type as a DataType. Once we've specialized dvarrays, we're going to need to include more information to raise the right error here. In this diff, we modify this IR op to take a StringData expected type; pre-specialization, this value will always be the current DataType string, but post-specialization, for array typehints, it'll be the specialized type.

Added a new test that checks that for builtins, post specialization, we correctly format both the expected and actual specialized types.

Reviewed By: jjgriego

Differential Revision: D21372161

fbshipit-source-id: 9bc21dddd05e101d7bf311d5a5e49869c3940d46