Post-specialization typehint enforcement
Summary:
Implement the post-specialization behavior rather than raising dvarray typehint notices. We force the notice flag on when specialization is on, so that HHBBC does the right "may throw" analysis of these bytecodes (the "may throw" logic is the same whether we're throwing notices or errors).

The core part of this change is to turn param, prop, and return typehint notices into type errors. That happens in four files:
- In type-constraint.cpp (interp regular functions), we just need to skip the notice and we'll fall through to the error.
- In irgen-types (JIT regular functions), we choose between genFail and dvArr.
- In native.cpp (interp builtin functions), we pull out the raise_type_error lambda and call it from the dvArr case.
- In irgen-builtin (JIT builtin functions), we call genFail from the dvArr case.

There are also a few other behaviors we have to change:
- In annot-type.h, we prevent escalation from ClsMeth to darray or to plain array.
- In class.cpp, when we create a subclass and the dvarrayness of a property doesn't match its superclass, we raise.
- In type-structure-helpers.cpp, we return false for "varray is shape" or "darray is tuple".

All new behaviors are tested in the specialization.php test. is/as and builtin type enforcement are easy to test. The param, property, and return type enforcement throw fatal errors, as does the subclass consistency check; we use APC to track how many times the test is run and test a single instance on each call to test_typehint_enforcement.

Reviewed By: jjgriego

Differential Revision: D21297508

fbshipit-source-id: 487ab82f5c7b33dd43c1e684c17caec4ed3ef087