Bad position for error relating to dynamic property access
Summary:
Error is reported at a bad position when accessing a property on a dynamic.

Note: there are other bugs lurking here, as the property access in the new test `dynamic_refine_then_invoke.php` should in fact be legal code. But because Hack sees the type of `$d` as the unsimplified type `((dynamic & nonnull) | (dynamic & null))`, it rejects an invocation on a `null` type.

But let's fix the position problem first.

Reviewed By: manzyuk

Differential Revision: D20870617

fbshipit-source-id: a1b9b078ed1e37c2945d88e6e3029570528f4cb8