Optimize idx/ArrayGet for OOB/undefined index on static arrays by keeping a side table
Summary:
This diff adds a side table attached to static ArrayData* in order to quickly check whether an element exists on the ArrayData*. Since make use of bottom bits of hashes, the table can give false positives but never a false negative.
We make use of ArrayAccessProfile to choose when to run this optimization.

Differential Revision: D21026058

fbshipit-source-id: d794cb74a535113088004ae5f8bef2fe779e1bd5