Emit side exit from array access profile when offset optimization fails with high profile threshold
Summary:
This diff makes it so that if ArrayAccessProfile tells us that a certain key will be at index i such that  98% of the samples were at that index, instead of emitting an unlikely (cold branch) fall back, it emits a sideexit. This sideexit, even though it makes the failure case more expensive, it reduces the amount of code needing to be emitted in the jit and removes phi points, eventually resulting in better optimized code.
This diff still maintains the emitting into cold when threshold is the previous threshold: 85.54%

Differential Revision: D21360412

fbshipit-source-id: 4297340110e0060c4df4de11a90ef945d9155e0d