Make namespacing behavior more consistent
Summary:
Prefix names in the global namespace with a backslash, as the OCaml decl parser does.

If we guarantee a trailing backslash on the current namespace name in the direct decl parser's NamespaceBuilder, it's a bit easier to qualify a name in the current namespace. We need only concatenate it to the current namespace (rather than checking to see if it ends in a backslash).

Reviewed By: Wilfred

Differential Revision: D21147055

fbshipit-source-id: d051720b62f6c3ffd23956411fffbb209de9aeec