Remove SmartConstructors::new method
Summary:
I'm working on arena-allocating the types produced by the direct decl parser. In order to do so, DirectDeclSmartConstructors needs to take an allocator as an input. It cannot construct and own the allocator itself, because allocated values borrow the allocator, which would make the DirectDeclSmartConstructors state self-referential.

Maybe we could try to work around this limitation with the `rental` crate, but I'd prefer not to. I think it's valuable to allow the caller to control the lifetime of the allocator. This would allow decl-parsing threads in the decl service to reuse the same allocator for many decl-parses (by [resetting](https://docs.rs/bumpalo/3.2.1/bumpalo/struct.Bump.html#method.reset) it between each parse, allowing the bump-allocator to retain some of the memory it grabbed from the global allocator, but conceptually "clearing" it).

Additionally, most of the implementations of this required trait method don't actually use all of the parameters, which is a bit of a code smell, I think. Better to let them each implement their own constructor functions which only take what's needed.

Reviewed By: Wilfred

Differential Revision: D21147217

fbshipit-source-id: 31e13ab8ea39166701cb22d50bac2d58edc63b57