Use oxidized_by_ref for decls and types
Summary:
In {D21147275}, I switched the direct decl parser over to use the oxidized_by_ref crate instead of the oxidized crate, allowing it to arena-allocate parsed decls.

To allow hh_check to continue using the direct decl parser, I needed to convert the by-ref decls into oxidized decls in that diff. But there is no particular reason why hh_check needs its DeclProvider to return oxidized decls. In fact, as long as the DeclProvider outlives the TypeBuilder and TAST (and typing_rust agrees with the decl provider on the representation of types), it is possible to directly reference types and positions from decls in inference types rather than cloning them (and this diff does so).

To get typing_rust and the direct decl parser to agree on the representation of types, I re-exported the oxidized_by_ref definitions in typing/typing_defs_core.rs, moved method implementations to oxidized_by_ref/manual/typing_defs_core_impl.rs, and revised hh_oxidize and oxidized_by_ref to more closely match manual definitions in typing_rust (notably, manually copying its changes to the representation of typing_reason). I stayed away from the practice of defining type aliases for reference types, e.g.,

```
pub struct FunParam_<'a> {...}
pub type FunParam<'a> = &'a FunParam_<'a>;
```

I did this partially to avoid implementation work to support the pattern in hh_oxidize, but also because I feel like our type definitions are easier to understand and more consistent if we avoid hiding references in this way (it is easier to tell which types are inlined into a containing struct or slice and which are not).

Because the AAST parser still produces the AST from the oxidized crate, we need to deal with two position types--the owned oxidized positions, and the borrowed oxidized_by_ref positions. When we wish to copy an AST position into a type, we need to clone the AST position into our arena (this diff adds `Pos::from_oxidized_in` for this purpose). Since switching to arena allocation in the direct-decl-parser resulted in a 20-25% performance improvement, perhaps in the long term, we could migrate the AAST parser and other tools to use oxidized_by_ref types instead, and eventually remove the oxidized crate.

Reviewed By: CatherineGasnier

Differential Revision: D21173898

fbshipit-source-id: dc20ef8d5fa2454327d5bd0c9bc7f746cc9f8bfd