Use make_namespace_declaration_header to drive NamespaceBuilder rather than make_token
Summary:
Because of the historical structure of the CST, the direct decl parser needed to implement some parsing logic in the `make_token` smart constructor, in order to track the current namespace and fully qualify names as they were constructed (rather than as a transformation after-the-fact). In {D20746539}, I changed the structure of the CST to resolve this problem (see the summary of that diff for a more detailed description of the problem and solution).

Now that the decl parser receives the name of the namespace before we actually enter it, we can get rid of the namespace-declaration-statement parsing logic in `make_token`, and set the current namespace in `make_namespace_declaration_header` instead.

Reviewed By: Wilfred

Differential Revision: D20840140

fbshipit-source-id: d14f36a090801e6700576cbd09cb4dd9eb156809