Remove dst from phpret
Summary:
The phpret vasm instruction has a dst argument, to indicate where it
will write the stored frame (optionally). However, this is only ever
set to rvmfp(). Moreover, this is a terminal instruction, so having a
dst is redundant, as no pass can make use of it.

This is harmless on its own, but I'm working on some register
allocation logic to track physical register lifetimes. As an
optimization, I can skip a lot of liveness analysis if a physical
register is never written to anywhere in the unit. This (redundant)
dst on phpret defeats that optimization.

So, remove the dst and just specify that the instruction writes to
rvmfp().

Reviewed By: ottoni

Differential Revision: D21415074

fbshipit-source-id: cab8d9bea286de7b61edb9f41d8afcd5f9ae3476