Add flag to disable func/string interop
Summary:
We're getting close to being able to drop support for func/string interop. This diff adds an option which disables interop.

To do this via a runtime option the diff also adds an alternate type for KindOfFunc which does not convert to string (and is not a member of the TStrLike union).

Reviewed By: jano

Differential Revision: D21384705

fbshipit-source-id: a13461539492b87a61466be9d9bcbeab57e5fc6b