State loader should return relative paths for dirty files
Summary:
It turns out that the 2 different saved state loaders we have make the opposite assumption about whether the list of changed files is a list of absolute paths or relative paths.

I believe that this is in part why the bug at the bottom of this stack was introduced (the other reason being the API having a confusing name).

This diff is a straightforward change of the type from string to Relative_path.t

I assumed that it's OK to treat all changed paths in the IDE/server as relative to root, since there are multiple places where we convert to Relative_path.t, anyway.

NOTE: An additional change could be made to add a non-dummy prefix type for those prefixes that are unknown in order to be able to handle type checking files that are not under the root, HHI, or temp paths.

Reviewed By: yinghuitan

Differential Revision: D21189618

fbshipit-source-id: 17eaa2aad49464966bddf6c4a761c6e2e1cd6612