Make state loader async
Summary:
I expect this change to be almost a refactor - only the next diff makes a difference in how the saved states are downloaded by ensuring they make progress in parallel.

This change just converts to async all the functions in State_loader that could be async.

Sorry about all the churn - it was about as fun to get all the types to unify when rewriting this as it is to read this diff.

Reviewed By: arxanas

Differential Revision: D21189621

fbshipit-source-id: d67807e299e9f7edff54735347194d56a77252d5