Future.make should support a timeout value
Summary:
This diff adds the ability to specify a timeout on the future itself and not just as a parameter to the Future's `get` function.

The purpose of this is to better support timeouts for futures that are bound continuations.

E.g., suppose you have a future chain like this:

```
A Future.t -> (A -> B Future.t) -> (B -> C Future.t)
```

After all the futures are composed, you get a single future that should yield the value of type C. You can call `Future.get c_future ~timeout:60`, but then you can't ensure that, say, no more than 5 seconds is spent on getting the A value - the 60 seconds are shared between the 3 futures.

The ultimate goal of this diff is to enable converting `State_loader` to be asynchronous, and since it's using specific timeout values for the individual steps that it consists of, I wanted to preserve that behavior.

Thus, this change: you can now either `make` the future with a timeout baked into it, or set the timeout on an existing future, which is convenient incase it's produced by another module that you don't want to plumb the timeout into.

Reviewed By: yinghuitan

Differential Revision: D21189620

fbshipit-source-id: b80974b7daa41015880ebc46c5533d6559e2ee76