Make the downloading of the 2 saved states execute in parallel
Summary:
To ensure that both saved states are getting downloaded in parallel, this change is about merging their futures and then polling the merged future until is is ready.

The underlying futures consist of chains of bound continuations, and, since our Future module does not have an executor mechanism beyond the caller polling the futures, that's what we have to do to make progress through all the continuations.

Previously, we had, say, the naming table future looking like this:

```
A Future.t (takes 5 seconds)
 -> B Future.T (takes 10 seconds)
  -> C Future.t (takes 20 seconds)
```

And we had this sequence of steps before:
1. Create the naming table future (which, through a sequence of steps, should eventually produce some value C)
2. Synchronously download the combined saved state (takes 30 seconds)
3. Wait for the naming table future to complete

The result was that we were waiting for 60 seconds for both saved states to finish downloading instead of 30 seconds.

The bottleneck was due to the fact that the naming table future only gets to produce value A on its own, and doesn't start on getting B and C until we're done getting the combined saved state.

After this diff stack, step (2) is also a chain of bound continuations, and both chains are now polled for readiness until they are both done.

Reviewed By: arxanas

Differential Revision: D21197353

fbshipit-source-id: 9d07f0e964c521e964ec2a25e79d7d3d43100ce6