Using PU syntax is only allowed in the directories specified by .hhconfig (supersedes D20516013)
Summary:
.hhconfig is extended with the following option:

      pocket_universe_enabled_paths = nowhere|everywhere|dir1, .., dirn

If the option is set to nowhere, or omitted, then using PU syntax is reported as a type error.  If it is set to dir1...dirn, the the PU syntax is enabled in directories dir1..dirn and their subdirectories.  If it is set to everywhere, the the use of PU syntax is unrestricted.

Directory paths dir1..dirn are relative to the location of the .hhconfig file.

In V2 I have added the option --enable-pocket-universes-syntax to hh_single_type_check (disabled by default).

Reviewed By: vsiles

Differential Revision: D21369102

fbshipit-source-id: 54346726b86555c8415d0bf21db5c09f2580c645