Add Node_::Ty variant
Summary:
The direct decl parser currently allocates an intermediate representation between AST hints and types called HintValue. This is currently necessary for some types. Since we do not know the names of a class's type parameters until we are finished parsing the class, we cannot know whether a type name is a generic parameter or not, and so we cannot decide whether to allocate a `Tapply` or `Tgeneric`. For other types, though, the IR doesn't add much. We can directly allocate types in those cases instead.

This diff adds the Node_::Ty variant for this purpose, and uses it for primitive types.

Reviewed By: Wilfred

Differential Revision: D21187228

fbshipit-source-id: 3b72b51c42ff903a01aa6c54cf1d7dd5e176056f