Track type parameters in scope using state
Summary:
The direct decl parser currently allocates an intermediate representation between AST hints and types called HintValue. This is because we did not collect class or function type parameters until constructing the decl for the class or function (i.e., *after* constructing all the inner HintValues). When converting HintValues to Tys, we would consider all the type parameters in scope to determine whether to allocate a Tapply or Tgeneric for a symbol in type position. This is why a set of type parameter names is threaded into functions like `node_to_ty`.

However, I don't think it's necessary for us to wait until we construct a class or function decl to make use of its type parameter list. We will parse the type parameter list before any of the inner types are parsed, and as a SmartConstructors implementation, we are informed of the type parameters as soon as they are parsed (via `make_type_parameters`). So instead, we can collect the type parameters in scope in our SmartConstructors state and eagerly construct Tys (avoiding the HintValue IR). When we are about to construct a class or function decl, we can pop a set of type parameters from our stack.

This diff does so, and (now that it is possible) removes HintValue::Varray to exercise the new behavior.

Reviewed By: Wilfred

Differential Revision: D21192276

fbshipit-source-id: c2d73e993c9b77b3415b8befc284638509eb4aba