Skip debugger tests consistenly and in more modes.
Summary:
The legacy debugger tests (using -m debug) don't work in any multi-request, server or repo modes, but we were only skipping them in repo mode and server modes (with 2 separate code paths).

Do skipping for all these modes in one place; this happens to be next to that for .verify, since the set of modes they don't work in is the same.

Use skip_test instead of can_run_server_test to filter configs we need to start servers for in server modes. Without this change we can't run `--server quick slow` in one go any more (limit is 30 configs, without the filtering changes we'd go from 29 to 31 configs). Because skip_test is more expensive we have to be more clever to avoid slowing down the filtering; we now first check if we already decided to use the given config for some other test before checking if the current test will run (thus requiring the config).

We also need to avoid a chicken-and-egg issue where we call skip_test to determine whether we need a server with this config, but skip_test needs a server to already be started to run a .skipif file against. Add an optional flag to skip_test that avoids running .skipif files, and pass this flag from the server starting code.

Reviewed By: oulgen

Differential Revision: D21401709

fbshipit-source-id: c1fcc795bc4da8c936a75f9a9b6e172eedf61227