Fix skip-count logic in run.php
Summary:
We can currently get a `--count` option from 3 sources: a multi_request_mode, from `extra_args($options)` in hhvm_cmd_impl and from `read_opts_file(find_test_ext($test, 'opts'))` in hhvm_cmd. If we end up with more than 1 source providing a `--count` option, the test can't work and needs to be skipped.

The logic doing this has a couple of problems: it only runs if we have a multi_request_mode, so running with `--args '--count=1'` causes tests that have `--count` in their .opts file fo fail.

Also, in `--cli-server` mode the command always has a `--count=3`, yet we skip if we see even one `--count` option in `--cli-server` mode. The only reason we don't skip all tests in this mode is that the regex looks for the count option with a space afterwards and the `--count=3` that `--cli-server` mode inserts usually ends up last in the command (but this is not the case if you have a `--jitsample` option set).

Fix this by looking for 2 `--count` options in all modes (without requiring a space after the second one), and in `--jit-serialize` mode (the only thing that adds a `--count` option after we run the check) checking for even one `--count` option (again without requiring a space after it).

Reviewed By: fredemmott

Differential Revision: D21401707

fbshipit-source-id: d5a284c4a14da19dc34eaa99f9a855b248df69b4