deal with exceptions during typechecking
Summary:
This diff regularizes D19501208, which introduced a top-level exception handler for typing_check_service.

(1) The comment in that diff was "[it's okay to use stderr] because everything printed to stderr goes to the log files". The story is that Hh_logger has two separate thresholds, one for stderr output and one for file output, and when you send a message to Hh_logger then it uses the right threshold. Moreover some binaries (e.g. clientLsp) simply swallow stderr and rely solely on Hh_logger writing to logfile. Thus, Hh_logger is the better abstraction, and we should be using it everywhere by default.

(2) I think that if we detect an error but fail to log it telemetry, then we might as well not have detected it in the first plae - since we'll have no idea of its prevalence in the wild.

(3) We were handling the exception backtrace correctly. But `Exception.t` is the nicer abstraction that attaches the backtrace to the exception. Rather than having two different abstractions floating around our codebase, we should be standardizing on Exception.t.

Reviewed By: CatherineGasnier

Differential Revision: D21403519

fbshipit-source-id: 5cbaec70fd1b993eb3072360110021a75219e478