Avoid constructing unused TokenStreams
Summary: When splitting OcamlRep into ToOcamlRep and FromOcamlRep, I did not change the `struct_impl` and `enum_impl` functions in ocamlrep_derive, which constructed TokenStreams for both the to_ocamlrep and from_ocamlrep function bodies. Instead, I just invoked them in each derive macro, and discarded the unused TokenStream. This is a little wasteful. With a minor refactor we can only construct the body that is needed.

Reviewed By: 2BitSalute

Differential Revision: D21182997

fbshipit-source-id: 76aa4ff12da55b9e32506945df9978f1f9143257