Fix bug in FromOcamlRep derive for boxed tuple variants
Summary:
In OCaml, variants with arguments are always represented with a pointer to a block. The variant tag is stored in the block header, and the block is exactly the size necessary to contain the arguments. This isn't true in Rust--enums are value types, and need to have a consistent size across all variants. This means that a Rust enum is the size of its largest variant (like a C union), plus the size of the tag, plus any necessary padding.

This means that a naive Rust representation of some of our OCaml types can be absurdly large, if one variant happens to have many (or large) arguments. The naive representation of `aast::Expr_`, for example, was 712 bytes! To deal with this, we opt in to putting variant fields behind an indirection in hh_oxidize for some types. This brought `aast::Expr_` down to just 32 bytes.

In {D17864959}, I added support for this pattern to ocamlrep_derive, but there was an oversight. When we are converting a boxed-tuple-variant OCaml value to Rust, the expected size of the OCaml block is the number of elements in the boxed tuple, not the number of arguments to the Rust variant (which is just one--the boxed tuple). Looks like I was missing test coverage for this. This diff fixes the bug and adds a test case.

Reviewed By: losvald

Differential Revision: D21182999

fbshipit-source-id: d90b5f06f2921fc1802b6eb0a776c4700c26057d