Make the test runner HAM-safe
Summary:
It's not valid to pass a varray to an array typehint after we turn on specialization.

In a couple of other places ($modes, $hhvm, $commands), we use a single string for a one-element list but a varray for a multiple-element list. Add a helper to coerce the string to the varray rather than doing array casts.

Reviewed By: markw65

Differential Revision: D21434124

fbshipit-source-id: 64fb95604e6501fdcd52d546529a15dfbe7d20ef