Reverse-naming-table-delta doesn't stores line+col
Summary:
We used to store `'pos pos_or_deleted` entries in the reverse-naming-table-delta, which amounts to this:
* always stores either `Full` position (file, line, column), or stores a `File` position (file, name_type)
  * name_type is one of Const, Fun, Class, RecordDef, TypeDef
* for classes+records+typedefs additionally stores `Naming_types.kind_of_type` which is one of Class, RecordDef, TypeDef

The delta is populated with (1) mostly when we look up names and fallback to sqlite then we store `File` positions, (2) also for files that change on disk since the saved-state then we store `Full` positions.

The `Full` position include line+col which are used in error reporting. If they're not there, then error reporting has to fetch them by looking up the file and parsing it.

In this diff:
1. I think that it's better to never store positions in memory. Errors are a rare case. Better to keep memory low, and retrieve positions if needed. I therefore removed the ability to store `Full` positions.
2. We therefore don't need to store `Naming_types.kind_of_type` ever, since equivalent information (name_type) was already being stored in some cases and is now being stored in all. This simplifies things.

Reviewed By: arxanas

Differential Revision: D21370550

fbshipit-source-id: 9aa04f6be0ff0a4f4d82633db0306d59fdb8fd2f