refactor ClientIdeIncremental
Summary:
In D20878723 we had to refactor ClientIdeIncremental. update_naming_tables_for_changed_file because most callers called it its promise-based version (it awaits a disk read), but hh_bulk_checker is like hh_server and so can't use Lwt and so needed a synchronous version.

That refactoring unfortunately introduced a lot of code duplication.

Truth be told, we're not actually benefiting from Lwt at all in this code path. None of the callers of this method have significant concurrent async workflows, and the latency of reading off disk is miniscule, so it's cleaner for us to switch wholly over to synchronous. Likewise, File_provider is entirely synchronous. It's strange to put effort into being asynchronous here in update_file when the much more critical code elsewhere is synchronous.

Reviewed By: arxanas

Differential Revision: D21367281

fbshipit-source-id: 21959fd9f547c9b2a3a7341121a700ef3b6d3817